check-status.service, online-service and user-service added/fixed)
check-status can count ALL stati, stati only from one instance, and check a status. online-service can count and check how many instances are online. user-serive is filled with dummy functions (needs additional work) which are supposed to check if a given name and password are correct and then save that this is a registered person, and can log this person out again